Move completion resolve handling logic into LSP service
