+++ src/Features/LanguageServer/Protocol/Handler/Completion/CompletionResolveHandler.cs
@@ -12,7 +12,6 @@ using Microsoft.CodeAnalysis.LanguageServer.Handler.Completion;
 using Microsoft.CodeAnalysis.LanguageService;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CommonLanguageServerProtocol.Framework;
-using Microsoft.VisualStudio.Text.Adornments;
 using Newtonsoft.Json.Linq;
 using Roslyn.Utilities;
 using LSP = Microsoft.VisualStudio.LanguageServer.Protocol;
@@ -21,10 +20,6 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler
 {
     /// <summary>
     /// Handle a completion resolve request to add description.
-    /// 
-    /// TODO - This must be moved to the MS.CA.LanguageServer.Protocol project once the
-    /// references to VS icon types and classified text runs are removed.
-    /// See https://github.com/dotnet/roslyn/issues/55142
     /// </summary>
     /// <remarks>
     /// This isn't a <see cref="ILspServiceDocumentRequestHandler{TRequest, TResponse}" /> because it could return null.
@@ -62,48 +57,24 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler
                 return completionItem;
             }
 
-            var list = cacheEntry.CompletionList;
-
             // Find the matching completion item in the completion list
-            var selectedItem = list.ItemsList.FirstOrDefault(cachedCompletionItem => MatchesLSPCompletionItem(completionItem, cachedCompletionItem));
-            if (selectedItem == null)
-            {
-                return completionItem;
-            }
+            var selectedItem = cacheEntry.CompletionList.ItemsList.FirstOrDefault(cachedCompletionItem => MatchesLSPCompletionItem(completionItem, cachedCompletionItem));
 
             var completionOptions = _globalOptions.GetCompletionOptions(document.Project.Language);
-            var displayOptions = _globalOptions.GetSymbolDescriptionOptions(document.Project.Language);
-            var description = await completionService.GetDescriptionAsync(document, selectedItem, completionOptions, displayOptions, cancellationToken).ConfigureAwait(false)!;
-            if (description != null)
+            var symbolDescriptionOptions = _globalOptions.GetSymbolDescriptionOptions(document.Project.Language);
+
+            if (selectedItem is not null)
             {
-                var supportsVSExtensions = clientCapabilities.HasVisualStudioLspCapability();
-                if (supportsVSExtensions)
-                {
-                    var vsCompletionItem = (LSP.VSInternalCompletionItem)completionItem;
-                    vsCompletionItem.Description = new ClassifiedTextElement(description.TaggedParts
-                        .Select(tp => new ClassifiedTextRun(tp.Tag.ToClassificationTypeName(), tp.Text)));
-                }
-                else
-                {
-                    var clientSupportsMarkdown = clientCapabilities.TextDocument?.Completion?.CompletionItem?.DocumentationFormat?.Contains(LSP.MarkupKind.Markdown) == true;
-                    completionItem.Documentation = ProtocolConversions.GetDocumentationMarkupContent(description.TaggedParts, document, clientSupportsMarkdown);
-                }
-            }
-
-            // We compute the TextEdit resolves for complex text edits (e.g. override and partial
-            // method completions) here. Lazily resolving TextEdits is technically a violation of
-            // the LSP spec, but is currently supported by the VS client anyway. Once the VS client
-            // adheres to the spec, this logic will need to change and VS will need to provide
-            // official support for TextEdit resolution in some form.
-            if (selectedItem.IsComplexTextEdit)
-            {
-                Contract.ThrowIfTrue(completionItem.InsertText != null);
-                Contract.ThrowIfTrue(completionItem.TextEdit != null);
-
-                var snippetsSupported = clientCapabilities?.TextDocument?.Completion?.CompletionItem?.SnippetSupport ?? false;
-
-                completionItem.TextEdit = await GenerateTextEditAsync(
-                    document, completionService, selectedItem, snippetsSupported, cancellationToken).ConfigureAwait(false);
+                var creationService = document.Project.Solution.Services.GetRequiredService<ILspCompletionResultCreationService>();
+                await creationService.ResolveAsync(
+                    completionItem,
+                    selectedItem,
+                    document,
+                    clientCapabilities,
+                    completionService,
+                    completionOptions,
+                    symbolDescriptionOptions,
+                    cancellationToken).ConfigureAwait(false);
             }
 
             return completionItem;
@@ -129,52 +100,6 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler
             // Now we're left with what should be the original display text for the lsp completion item.
             // Check to make sure it matches the cached completion item label.
             return string.Equals(originalDisplayText, completionItem.DisplayText);
-        }
-
-        // Internal for testing
-        internal static async Task<LSP.TextEdit> GenerateTextEditAsync(
-            Document document,
-            CompletionService completionService,
-            CompletionItem selectedItem,
-            bool snippetsSupported,
-            CancellationToken cancellationToken)
-        {
-            var documentText = await document.GetTextAsync(cancellationToken).ConfigureAwait(false);
-
-            var completionChange = await completionService.GetChangeAsync(
-                document, selectedItem, cancellationToken: cancellationToken).ConfigureAwait(false);
-            var completionChangeSpan = completionChange.TextChange.Span;
-            var newText = completionChange.TextChange.NewText;
-            Contract.ThrowIfNull(newText);
-
-            // If snippets are supported, that means we can move the caret (represented by $0) to
-            // a new location.
-            if (snippetsSupported)
-            {
-                var caretPosition = completionChange.NewPosition;
-                if (caretPosition.HasValue)
-                {
-                    // caretPosition is the absolute position of the caret in the document.
-                    // We want the position relative to the start of the snippet.
-                    var relativeCaretPosition = caretPosition.Value - completionChangeSpan.Start;
-
-                    // The caret could technically be placed outside the bounds of the text
-                    // being inserted. This situation is currently unsupported in LSP, so in
-                    // these cases we won't move the caret.
-                    if (relativeCaretPosition >= 0 && relativeCaretPosition <= newText.Length)
-                    {
-                        newText = newText.Insert(relativeCaretPosition, "$0");
-                    }
-                }
-            }
-
-            var textEdit = new LSP.TextEdit()
-            {
-                NewText = newText,
-                Range = ProtocolConversions.TextSpanToRange(completionChangeSpan, documentText),
-            };
-
-            return textEdit;
         }
 
         private CompletionListCache.CacheEntry? GetCompletionListCacheEntry(LSP.CompletionItem request)
